requirements.txt
pandas~=1.0.3
notebook
scikit-learn~=0.22.2.post1
keras
maya~=0.6.1
numpy~=1.18.2
treelib~=1.6.1
tensorflow-gpu~=2.1.0